pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pk-agent"
version = "0.1.0"
description = "Screen capture personal KB agent; vision gate (Haiku) + hint generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.22",
    "httpx>=0.27.0",
    "mss>=9.0.1",
    "numpy>=1.26.0",
    "anthropic>=0.40.0",
    "pillow>=10.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.12.0",
]

[project.scripts]
pk-agent = "pk_agent.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pk_agent"]

[tool.hatch.build.targets.wheel.sources]
"src/pk_agent" = "pk_agent"

[tool.hatch.build.targets.sdist]
include = ["src/pk_agent"]
